+++ tests/apps/dummy/token.py
@@ -1,13 +1,13 @@
 from starlette.authentication import SimpleUser
 
-from starlette_auth_toolkit.base.backends import BearerAuthBackend
+from starlette_auth_toolkit.base.backends import BaseTokenAuth
 
 from ..utils import get_base_app
 
 TOKEN = "s3kr3t_t0k3n"
 
 
-class BearerAuth(BearerAuthBackend):
+class TokenAuth(BaseTokenAuth):
     async def verify(self, token: str):
         if token != TOKEN:
             return None
@@ -15,4 +15,4 @@ class BearerAuth(BearerAuthBackend):
 
 
 def get_app():
-    return get_base_app(backend=BearerAuth())
+    return get_base_app(backend=TokenAuth())