+++ tests/test_auth_token.py
@@ -1,7 +1,7 @@
 import pytest
 from starlette.testclient import TestClient
 
-from .apps.dummy.bearer import TOKEN, get_app
+from .apps.dummy.token import TOKEN, get_app
 
 
 @pytest.fixture(name="client")
@@ -12,10 +12,10 @@ def fixture_client():
 @pytest.mark.parametrize(
     "token, status_code", [(TOKEN, 200), (None, 403), ("doesnotexist", 401)]
 )
-def test_bearer_auth(client, token, status_code):
+def test_auth(client, token, status_code):
     headers = {}
     if token:
-        headers["Authorization"] = f"Bearer {token}"
+        headers["Authorization"] = f"Token {token}"
 
     r = client.get("/", headers=headers)
     assert r.status_code == status_code